[ANDROSDK-2009] move api logic to network layer
